+++ apps/client/package.json
@@ -34,11 +34,13 @@
     "graphql": "^16.5.0",
     "ipfs-http-client": "^59.0.0",
     "lens": "*",
+    "mediasoup-client": "^3.6.68",
     "nanoid": "^4.0.0",
     "next": "13.1.1",
     "next-auth": "4.18.6",
     "next-axiom": "^0.16.0",
     "next-pwa": "^5.6.0",
+    "protocol": "*",
     "react": "^18.2.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
