+++ apps/client/src/app/helpers/quaternionToYaw.ts
@@ -1,0 +1,4 @@
+export function quaternionToYaw(y: number, w: number): number {
+  const yaw = Math.atan2(2 * y * w, 1 - 2 * y * y);
+  return yaw;
+}
