--- apps/client/src/app/networking/NetworkingInterface.ts
@@ -1,497 +1,0 @@
-import { nanoid } from "nanoid";
-import { BehaviorSubject } from "rxjs";
-
-import { FromHostMessage, InternalChatMessage, SpaceJoinStatus, ToHostMessage } from "./types";
-import { WebRTC } from "./WebRTC";
-
-/*
- * Acts as an interface for all networking functionality.
- * Handles joining spaces, sending and receiving messages, etc.
- */
-export class NetworkingInterface {
-  #webRTC: WebRTC | null = null;
-  #ws: WebSocket | null = null;
-
-  #spaceNodeId: string | null = null;
-  #broadcastInterval: NodeJS.Timeout | null = null;
-  #hostServer: string | null = null;
-  #reconnectCount = 0;
-  #producedTrack: MediaStreamTrack | null = null;
-
-  playerPosition: Int32Array | null = null;
-  playerRotation: Int16Array | null = null;
-  cameraPosition: Int32Array | null = null;
-  cameraRotation: Int16Array | null = null;
-
-  #connectedPlayers = new Set<number>([-1]);
-  #loadedPlayers = new Set<number>();
-  #playerNames = new Map<number, string>();
-  #playerAddresses = new Map<number, string>();
-
-  #myName: string | null = null;
-  #myAvatar: string | null = null;
-  #myAddress: string | null = null;
-
-  playerId$ = new BehaviorSubject<number | null>(null);
-  chatMessages$ = new BehaviorSubject<InternalChatMessage[]>([]);
-
-  spaceJoinStatus$ = new BehaviorSubject<SpaceJoinStatus>({
-    spaceId: null,
-    wsConnected: false,
-    webrtcConnected: false,
-    sceneLoaded: false,
-  });
-
-  get spaceJoinStatus() {
-    return this.spaceJoinStatus$.value;
-  }
-
-  set spaceJoinStatus(status: SpaceJoinStatus) {
-    this.spaceJoinStatus$.next(status);
-  }
-
-  constructor() {}
-
-  async joinSpace({
-    spaceId,
-    host,
-    modelURL,
-  }: {
-    spaceId: number;
-    host: string;
-    modelURL: string;
-  }) {
-    this.#reconnectCount = 0;
-
-    this.spaceJoinStatus = {
-      spaceId,
-      wsConnected: false,
-      webrtcConnected: false,
-      sceneLoaded: false,
-    };
-
-    this.#hostServer = host;
-
-    // Connect to host server
-    this.#connectToHost(spaceId);
-
-    this.spaceJoinStatus = {
-      ...this.spaceJoinStatus,
-      sceneLoaded: true,
-    };
-
-    // Wait for space to load
-    await new Promise((resolve, reject) => {
-      const interval = setInterval(() => {
-        if (this.spaceJoinStatus.spaceId !== spaceId) {
-          clearInterval(interval);
-          reject();
-          return;
-        }
-
-        if (
-          this.spaceJoinStatus.wsConnected &&
-          this.spaceJoinStatus.webrtcConnected &&
-          this.#loadedPlayers.size === this.#connectedPlayers.size
-        ) {
-          clearInterval(interval);
-          resolve(null);
-        }
-      }, 100);
-    });
-  }
-
-  #connectToHost(spaceId: number) {
-    if (!this.#hostServer) throw new Error("Host server not set");
-
-    // Create WebSocket connection to host server
-    const ws = new WebSocket(this.#hostServer);
-    this.#ws = ws;
-
-    // Create WebRTC manager
-    this.#webRTC = new WebRTC(ws, this, this.#renderThread, this.#producedTrack);
-    this.#webRTC.playerId = this.playerId$.value;
-    this.#webRTC.playerPosition = this.playerPosition;
-    this.#webRTC.playerRotation = this.playerRotation;
-    this.#webRTC.cameraPosition = this.cameraPosition;
-    this.#webRTC.cameraRotation = this.cameraRotation;
-
-    function send(message: ToHostMessage) {
-      ws.send(JSON.stringify(message));
-    }
-
-    ws.onopen = () => {
-      console.info("✅ Connected to host");
-      this.#reconnectCount = 0;
-
-      this.spaceJoinStatus = {
-        ...this.spaceJoinStatus,
-        wsConnected: true,
-      };
-
-      // Send player info
-      if (this.#myAvatar) this.#sendAvatar();
-      if (this.#myAddress) this.#sendAddress();
-      if (this.#myName) this.#sendName();
-
-      // Start WebRTC connection
-      if (!this.#webRTC) throw new Error("WebRTC not initialized");
-      this.#webRTC.connect();
-
-      // Join space
-      send({ subject: "join", data: { id: spaceId } });
-    };
-
-    ws.onmessage = (event: MessageEvent<string>) => {
-      this.#webRTC?.onmessage(JSON.parse(event.data));
-
-      const { subject, data }: FromHostMessage = JSON.parse(event.data);
-
-      switch (subject) {
-        case "join_successful": {
-          // Set your name
-          if (this.#myName) this.#playerNames.set(data.playerId, this.#myName);
-          else this.#playerNames.delete(data.playerId);
-
-          // Set your address
-          if (this.#myAddress) this.#playerAddresses.set(data.playerId, this.#myAddress);
-          else this.#playerAddresses.delete(data.playerId);
-
-          // Save player id
-          this.playerId$.next(data.playerId);
-          if (this.#webRTC) this.#webRTC.playerId = data.playerId;
-
-          break;
-        }
-
-        case "player_joined": {
-          console.info(`👋 Player ${toHex(data.playerId)} joined`);
-
-          this.#connectedPlayers.add(data.playerId);
-
-          // Set name
-          if (data.name) this.#playerNames.set(data.playerId, data.name);
-          else this.#playerNames.delete(data.playerId);
-
-          // Set address
-          if (data.address) this.#playerAddresses.set(data.playerId, data.address);
-          else this.#playerAddresses.delete(data.playerId);
-
-          // Add player to scene
-          this.#renderThread.postMessage({ subject: "player_joined", data });
-
-          // Get player name
-          const username = this.#getUsername(data.playerId);
-
-          this.#renderThread.postMessage({
-            subject: "player_name",
-            data: {
-              playerId: data.playerId,
-              name: username,
-            },
-          });
-
-          // Add message to chat if they joined after you
-          if (!data.beforeYou)
-            this.#addChatMessage({
-              type: "system",
-              variant: "player_joined",
-              id: nanoid(),
-              timestamp: Date.now(),
-              playerId: data.playerId,
-              username,
-            });
-
-          break;
-        }
-
-        case "player_left": {
-          console.info(`👋 Player ${toHex(data)} left`);
-
-          this.#connectedPlayers.delete(data);
-          this.#loadedPlayers.delete(data);
-
-          // Delete name
-          this.#playerNames.delete(data);
-
-          this.#renderThread.postMessage({ subject: "player_left", data });
-
-          // Get player name
-          const username = this.#getUsername(data);
-
-          // Add message to chat
-          this.#addChatMessage({
-            type: "system",
-            variant: "player_left",
-            id: nanoid(),
-            timestamp: Date.now(),
-            playerId: data,
-            username,
-          });
-          break;
-        }
-
-        case "player_message": {
-          // Get player name
-          const username = this.#getUsername(data.playerId);
-
-          // Add message to chat
-          this.#addChatMessage({
-            type: "chat",
-            username,
-            ...data,
-          });
-          break;
-        }
-
-        case "player_falling_state": {
-          this.#renderThread.postMessage({
-            subject: "set_player_falling_state",
-            data,
-          });
-          break;
-        }
-
-        case "player_name": {
-          console.info(`📇 Player ${toHex(data.playerId)} is now ${data.name}`);
-
-          if (data.name) this.#playerNames.set(data.playerId, data.name);
-          else this.#playerNames.delete(data.playerId);
-
-          const username = this.#getUsername(data.playerId);
-
-          this.#renderThread.postMessage({
-            subject: "player_name",
-            data: {
-              playerId: data.playerId,
-              name: username,
-            },
-          });
-          break;
-        }
-
-        case "player_avatar": {
-          console.info(`💃 Got custom avatar for ${toHex(data.playerId)}`);
-
-          this.#loadedPlayers.delete(data.playerId);
-
-          this.#renderThread.postMessage({
-            subject: "set_player_avatar",
-            data,
-          });
-          break;
-        }
-
-        case "player_address": {
-          console.info(`🦩 Player ${toHex(data.playerId)} is now ${data.address}`);
-
-          if (data.address) this.#playerAddresses.set(data.playerId, data.address);
-          else this.#playerAddresses.delete(data.playerId);
-
-          const username = this.#getUsername(data.playerId);
-
-          this.#renderThread.postMessage({
-            subject: "player_name",
-            data: {
-              playerId: data.playerId,
-              name: username,
-            },
-          });
-          break;
-        }
-      }
-    };
-
-    ws.onclose = async () => {
-      console.info("❌ Disconnected from host");
-
-      // Remove all players from scene
-      this.#playerAddresses.clear();
-      this.#playerNames.clear();
-      this.#renderThread.postMessage({ subject: "clear_players", data: null });
-
-      if (this.#broadcastInterval) clearInterval(this.#broadcastInterval);
-
-      if (!this.#hostServer || this.#reconnectCount > 0) return;
-
-      // Try reconnect
-      while (this.#reconnectCount < 10) {
-        const count = ++this.#reconnectCount;
-
-        // Wait a little longer each attempt
-        const timeoutLength = Math.min(1000 * count);
-        await new Promise((resolve) => setTimeout(resolve, timeoutLength));
-
-        // Test if has been reconnected
-        if (this.#ws?.readyState === WebSocket.OPEN) return;
-
-        // Close preview WebSocket connection
-        if (this.#ws) this.#ws.close();
-        this.#ws = null;
-
-        // If host server has been removed, give up
-        if (!this.#hostServer) return;
-
-        console.info(`🔄 (${count}) Attempting reconnect to host...`);
-        this.#connectToHost(spaceId);
-      }
-
-      console.error("🪦 Failed to reconnect to host. Giving up.");
-      this.disconnect();
-    };
-  }
-
-  disconnect() {
-    // Close WebRTC connection
-    if (this.#webRTC) this.#webRTC.disconnect();
-
-    // Close WebSocket connection
-    if (this.#ws) this.#ws.close();
-    this.#ws = null;
-
-    this.#reconnectCount = 0;
-    this.#hostServer = null;
-    if (this.#broadcastInterval) clearInterval(this.#broadcastInterval);
-  }
-
-  leaveSpace() {
-    this.disconnect();
-
-    // Remove space node from scene
-    if (this.#spaceNodeId) {
-      this.#scene.removeNode(this.#spaceNodeId);
-      this.#spaceNodeId = null;
-    }
-  }
-
-  sendChatMessage(data: string) {
-    if (!this.#ws?.OPEN) return;
-
-    const message: ToHostMessage = {
-      subject: "message",
-      data,
-    };
-
-    this.#ws.send(JSON.stringify(message));
-  }
-
-  setPlayerPosition(position: Int32Array) {
-    this.playerPosition = position;
-    if (this.#webRTC) this.#webRTC.playerPosition = position;
-  }
-
-  setPlayerRotation(rotation: Int16Array) {
-    this.playerRotation = rotation;
-    if (this.#webRTC) this.#webRTC.playerRotation = rotation;
-  }
-
-  setCameraPosition(position: Int32Array) {
-    this.cameraPosition = position;
-    if (this.#webRTC) this.#webRTC.cameraPosition = position;
-  }
-
-  setCameraRotation(rotation: Int16Array) {
-    this.cameraRotation = rotation;
-    if (this.#webRTC) this.#webRTC.cameraRotation = rotation;
-  }
-
-  setFallState(falling: boolean) {
-    if (!this.#ws || !this.#isWsOpen()) return;
-
-    const message: ToHostMessage = {
-      subject: "falling_state",
-      data: falling,
-    };
-
-    this.#ws.send(JSON.stringify(message));
-  }
-
-  setName(name: string | null) {
-    this.#myName = name;
-    this.#sendName();
-  }
-
-  #sendName() {
-    if (!this.#ws || !this.#isWsOpen()) return;
-
-    const message: ToHostMessage = {
-      subject: "set_name",
-      data: this.#myName,
-    };
-
-    this.#ws.send(JSON.stringify(message));
-  }
-
-  setAvatar(url: string | null) {
-    this.#myAvatar = url;
-    this.#renderThread.postMessage({ subject: "set_avatar", data: url });
-    this.#sendAvatar();
-  }
-
-  #sendAvatar() {
-    if (!this.#ws || !this.#isWsOpen()) return;
-
-    const message: ToHostMessage = {
-      subject: "set_avatar",
-      data: this.#myAvatar,
-    };
-
-    this.#ws.send(JSON.stringify(message));
-  }
-
-  setAddress(address: string | null) {
-    this.#myAddress = address;
-    this.#sendAddress();
-  }
-
-  #sendAddress() {
-    if (!this.#ws || !this.#isWsOpen()) return;
-
-    const message: ToHostMessage = {
-      subject: "set_address",
-      data: this.#myAddress,
-    };
-
-    this.#ws.send(JSON.stringify(message));
-  }
-
-  #addChatMessage(message: InternalChatMessage) {
-    const newChatMessages = this.chatMessages$.value.concat(message);
-
-    // Sort by timestamp
-    newChatMessages.sort((a, b) => b.timestamp - a.timestamp);
-
-    // Limit to 50 messages
-    newChatMessages.splice(50, newChatMessages.length - 50);
-
-    this.chatMessages$.next(newChatMessages);
-  }
-
-  #getUsername(playerId: number) {
-    const address = this.#playerAddresses.get(playerId);
-    const name = this.#playerNames.get(playerId);
-
-    const username = address ? address.substring(0, 6) : name ?? `Guest ${toHex(playerId)}`;
-
-    return username;
-  }
-
-  #isWsOpen() {
-    return this.#ws && this.#ws.readyState === this.#ws.OPEN;
-  }
-
-  produceAudio(track: MediaStreamTrack) {
-    if (!this.#webRTC) throw new Error("WebRTC not initialized");
-    this.#producedTrack = track;
-    return this.#webRTC.produceAudio(track);
-  }
-
-  setAudioPaused(paused: boolean) {
-    if (!this.#webRTC) throw new Error("WebRTC not initialized");
-    this.#webRTC.setAudioPaused(paused);
-  }
-
-  setPlayerLoaded(playerId: number) {
-    if (this.#connectedPlayers.has(playerId)) this.#loadedPlayers.add(playerId);
-  }
-}
