--- apps/client/src/app/networking/constants.ts
@@ -1,1 +1,0 @@
-export const LENS_API = "https://api-mumbai.lens.dev/";
