+++ apps/host/package.json
@@ -16,9 +16,9 @@
     "start": "node --experimental-specifier-resolution=node ./build/server.js"
   },
   "dependencies": {
-    "engine": "*",
     "mediasoup": "^3.11.4",
     "nanoid": "^4.0.0",
+    "protocol": "*",
     "uWebSockets.js": "uNetworking/uWebSockets.js#v20.15.0",
     "zod": "^3.20.2"
   },
