+++ apps/host/src/utils/send.ts
@@ -1,4 +1,4 @@
-import { FromHostMessage } from "engine";
+import { FromHostMessage } from "protocol";
 import uWS from "uWebSockets.js";
 
 export function send(ws: uWS.WebSocket, message: FromHostMessage) {
