+++ packages/contracts/README.md
@@ -1,3 +1,3 @@
-# contracts
+# Contracts
 
 Contains generated types for contracts used by the Wired, using [TypeChain](https://github.com/dethcrypto/TypeChain).
