+++ packages/engine/package.json
@@ -15,7 +15,6 @@
     "@gltf-transform/functions": "^2.4.7",
     "@pixiv/three-vrm": "^1.0.5",
     "contracts": "*",
-    "mediasoup-client": "^3.6.68",
     "meshoptimizer": "^0.18.1",
     "nanoid": "^4.0.0",
     "rxjs": "^7.8.0",
@@ -25,8 +24,6 @@
   "devDependencies": {
     "@types/offscreencanvas": "^2019.7.0",
     "@types/three": "^0.144.0",
-    "eslint": "^8.30.0",
-    "eslint-config-custom": "*",
     "tsconfig": "*",
     "typescript": "~4.9.4"
   }
