+++ packages/engine/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./Engine";
 export * from "./main/MainScene";
-export * from "../../../apps/client/src/app/networking/types";
 export * from "./scene";
 export * from "./types";
