+++ packages/protocol/README.md
@@ -1,0 +1,3 @@
+# Protocol
+
+Contains types describing the interface between client and host.
