+++ packages/protocol/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "protocol",
+  "version": "0.0.0",
+  "private": true,
+  "license": "GPL-3.0-or-later",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "scripts": {
+    "lint": "cd ../.. && eslint --fix packages/protocol --ext .js,.cjs,.mjs,.jsx,.ts,.tsx,.json --ignore-path .prettierignore"
+  },
+  "devDependencies": {
+    "tsconfig": "*",
+    "typescript": "~4.9.4",
+    "mediasoup-client": "^3.6.68"
+  }
+}
