+++ packages/protocol/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "rootDir": "./src"
+  },
+  "extends": "tsconfig/base.json",
+  "include": ["src"]
+}
